i64 index fields
Source: https://github.com/msupply-foundation/open-msupply/commit/e8af2b44868af84dbaedbf1d989504c02818abe7

+++ server/repository/src/migrations/v2_04_00/indicator_column_create_table.rs
@@ -14,7 +14,7 @@ impl MigrationFragment for Migrate {
             CREATE TABLE indicator_column (
                 id TEXT PRIMARY KEY NOT NULL,
                 program_indicator_id TEXT NOT NULL REFERENCES program_indicator(id),
-                index INTEGER NOT NULL,
+                index BIGINT NOT NULL,
                 header TEXT NOT NULL,
                 value_type TEXT NOT NULL,
                 default_value TEXT NOT NULL,
